server/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
python-dotenv>=1.0.1
redis>=5.0.0
./custom_packages/agenthub_sdk-0.1.4-py3-none-any.whl
